Make STM32 9-bit UART somewhat more functional.
The size argument to `HAL_UART_Receive_IT` and `HAL_UART_Transmit_IT` is the number of frames.
In 9-bit mode, it effectively casts the `uint8_t *` data argument to `uint16_t *`.

> When UART parity is not enabled (PCE = 0), and Word Length is configured to 9 bits (M1-M0 = 01),
> the received data is handled as a set of u16. In this case, Size must indicate the number
> of u16 available through pData.

On receive, use a two-long buffer so receive both octets. Then copy both into the Python destination.
On transmit, halve the octet count to get the size.

This does not address any of the subtantive issues in issue 2354.
A caller can use `array('H'`, but must expect the available / received / transmitted counts to be doubled,
as they still represent bytes moved by Python.